fix some stuff broken by rebase
--HG--
extra : rebase_source : 91bc42b90de20f91a8c5ac3eb1f8d5df758e5c7c

+++ src/java/com/acunu/castle/control/ArrayInfo.java
@@ -211,12 +211,12 @@ public class ArrayInfo extends DAObject {
 		return currentSizeInBytes;
 	}
 
-	public void setValueExIds(SortedSet<Integer> valueExIds)
+	public void setValueExIds(SortedSet<Long> valueExIds)
 	{
 		this.valueExIds = valueExIds;
 	}
 
-	public SortedSet<Integer> getValueExIds()
+	public SortedSet<Long> getValueExIds()
 	{
 		return valueExIds;
 	}
